Refactor SignUp view.
This commit refactors the SignUp view to inline functionality in the view, rather than in a separate component.
